requirements.txt
python-telegram-bot==22.5
openai==2.7.1
anthropic==0.31.2
Pillow==10.4.0